Update QuadForm.h
signed to unsigned

+++ src/QuadForm.h
@@ -804,9 +804,9 @@ public:
         os << "QuadForm(" << q.a_ << "," << q.b_ << "," << q.c_ << ","
             << q.f_ << "," << q.g_ << "," << q.h_ << ")";
       */
-        for (int i = 0; i < Rank; i++)
+        for (size_t i = 0; i < Rank; i++)
 	  {
-	    for (int j = 0; j < Rank; j++)
+	    for (size_t j = 0; j < Rank; j++)
 	      std::cout << q.B_[i][j] << " ";
 	    std::cout << std::endl;
 	  }
